% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignDefinitive.R
\name{assignDefinitive}
\alias{assignDefinitive}
\title{assignDefinitive}
\usage{
assignDefinitive(df = NULL, masterList = NULL)
}
\arguments{
\item{df}{This is a dataframe of species for which we have no definitive code}

\item{masterList}{This is the originally submitted list of species (with an 
internally assigned ID)}
}
\description{
This function takes a dataframe of results from a webservice, 
and determines whether the newly found values are definitive, or if they 
should be sent to additional web services in an attempt to get a more 
conclusive result.
If multiple services return the same code, the results are aggregated so that 
all of the contributing services will be captured (e.g. this code was the 
result of taxize AND worrms)
}
\seealso{
Other speciesCodes: \code{\link{cleanPrepareSpecList}},
  \code{\link{do_ritis}}, \code{\link{do_taxize}},
  \code{\link{do_worrmsTSN}}, \code{\link{do_worrms}},
  \code{\link{getTaxaIDs}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
