% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_ritis.R
\name{chk_ritis}
\alias{chk_ritis}
\title{chk_ritis}
\usage{
chk_ritis(recs = NULL, searchtype = NULL)
}
\arguments{
\item{recs}{- vector of taxa names for which the TSN is either unknown or not
definitive}

\item{searchtype}{- flag indicating whether scientific or common names should 
be used checking the services}
}
\description{
This function sends either the scientific or common name of each
species off to ritis
}
\seealso{
Other speciesCodes: \code{\link{assignDefinitive}},
  \code{\link{chk_taxize}}, \code{\link{chk_worrmsTSN}},
  \code{\link{chk_worrms}}, \code{\link{getAphiaIDs}},
  \code{\link{getTSNs}}, \code{\link{getTaxaIDs}},
  \code{\link{skipUselessRecs}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
