% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_taxize.R
\name{chk_taxize}
\alias{chk_taxize}
\title{chk_taxize}
\usage{
chk_taxize(df = NULL, field = NULL, searchtype = NULL)
}
\arguments{
\item{df}{- df of species for which the APHIAID is either unknown or not
definitive}

\item{field}{- field in the df containing the value to check against the 
service}

\item{searchtype}{- flag indicating whether scientific or common names should 
be used checking the services}
}
\description{
This function sends either the scientific or common name of each
species off to taxize
}
\seealso{
Other speciesCodes: \code{\link{assignDefinitive}},
  \code{\link{chk_ritis}}, \code{\link{chk_worrmsTSN}},
  \code{\link{chk_worrms}}, \code{\link{getAphiaIDs}},
  \code{\link{getTSNs}}, \code{\link{getTaxaIDs}},
  \code{\link{skipUselessRecs}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
