% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_worrms.R
\name{chk_worrms}
\alias{chk_worrms}
\title{chk_worrms}
\usage{
chk_worrms(recs = NULL, searchtype = NULL)
}
\arguments{
\item{recs}{-  - vector of taxa names for which the APHIAID is either unknown or not
definitive}

\item{searchtype}{- flag indicating whether scientific or common names should 
be used checking the services}
}
\description{
This function sends either the scientific or common name of each
species off to worrms
}
\seealso{
Other speciesCodes: \code{\link{assignDefinitive}},
  \code{\link{chk_ritis}}, \code{\link{chk_taxize}},
  \code{\link{chk_worrmsTSN}}, \code{\link{getAphiaIDs}},
  \code{\link{getTSNs}}, \code{\link{getTaxaIDs}},
  \code{\link{skipUselessRecs}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
