% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_worrmsTSN.R
\name{chk_worrmsTSN}
\alias{chk_worrmsTSN}
\title{chk_worrmsTSN}
\usage{
chk_worrmsTSN(recs = NULL, knownAphias = NULL)
}
\arguments{
\item{recs}{- vector of taxa names for which we already have an APHIAID}

\item{knownAphias}{- dataframe of taxas for which we have the AphiaIDs and a
field called "SCI_COL_CLN" containing and identifying name for the taxa.}
}
\description{
This function sends APHIAIDs to worrms to see if it can find a 
corresponding TSN
}
\seealso{
Other speciesCodes: \code{\link{assignDefinitive}},
  \code{\link{chk_ritis}}, \code{\link{chk_taxize}},
  \code{\link{chk_worrms}}, \code{\link{getAphiaIDs}},
  \code{\link{getTSNs}}, \code{\link{getTaxaIDs}},
  \code{\link{skipUselessRecs}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
