% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_sp.r
\name{df_to_sp}
\alias{df_to_sp}
\title{df_to_sp}
\usage{
df_to_sp(df = NULL, lat.field = "LATITUDE", lon.field = "LONGITUDE",
  the.CRS = "+init=epsg:4326")
}
\arguments{
\item{df}{default is \code{NULL}. This is the dataframe you want to spatialize.}

\item{lat.field}{the default is \code{"LATITUDE"}. the name of the field holding latitude values
(in decimal degrees)}

\item{lon.field}{the default is \code{"LONGITUDE"}.  the name of the field holding longitude
values (in decimal degrees)}

\item{the.CRS}{the default is \code{"+init=epsg:4326"}. This identifies the projection for the input data}
}
\value{
spatialpointsdataframe
}
\description{
This function takes a dataframe and converts in to a spatialpointsdataframe
}
\seealso{
Other data: \code{\link{make_oracle_cxn}},
  \code{\link{serviceMaker}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
