% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ritis.R
\name{do_ritis}
\alias{do_ritis}
\title{do_ritis}
\usage{
do_ritis(df = NULL, chkField = NULL, logName = logName,
  searchtype = NULL)
}
\arguments{
\item{df}{- dataframe with names you want to check against ritis}

\item{chkField}{- name of the field in the df you want to check}

\item{logName}{- this is the name of the logfile in the working directory 
that progress should be appended to.}

\item{searchtype}{- flag indicating whether scientific or common names should 
be used checking the services}
}
\description{
This function sends either the scientific or common name of each
species off to ritis
}
\seealso{
Other speciesCodes: \code{\link{assignDefinitive}},
  \code{\link{cleanPrepareSpecList}},
  \code{\link{do_taxize}}, \code{\link{do_worrmsAphiaID}},
  \code{\link{do_worrmsTSN}}, \code{\link{do_worrms}},
  \code{\link{getTaxaIDs}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
