% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_worrmsAphiaID.R
\name{do_worrmsAphiaID}
\alias{do_worrmsAphiaID}
\title{do_worrmsAphiaID}
\usage{
do_worrmsAphiaID(recs = NULL, knownTSNs = NULL, logName = logName)
}
\arguments{
\item{recs}{- vector of taxa names for which we already have a TSN}

\item{knownTSNs}{- dataframe of taxas for which we have the TSNs and a
field called "SCI_COL_CLN" containing and identifying name for the taxa.}

\item{logName}{- this is the name of the logfile in the working directory 
that progress should be appended to.}
}
\description{
This function sends TSNS to worrms to see if it can find a 
corresponding APHIAID
}
\seealso{
Other speciesCodes: \code{\link{assignDefinitive}},
  \code{\link{cleanPrepareSpecList}},
  \code{\link{do_ritis}}, \code{\link{do_taxize}},
  \code{\link{do_worrmsTSN}}, \code{\link{do_worrms}},
  \code{\link{getTaxaIDs}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
