% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAphiaIDs.R
\name{getAphiaIDs}
\alias{getAphiaIDs}
\title{getAphiaIDs}
\usage{
getAphiaIDs(mysteryAPHIAID = NULL, doSci = T, doComm = T,
  masterList = NULL)
}
\arguments{
\item{mysteryAPHIAID}{- df of species for which the APHIAID is either unknown or not
definitive}

\item{doSci}{- flag indicating whether scientific names should be used when 
checking the services}

\item{doComm}{- flag indicating whether common names should be used when 
checking the services}

\item{masterList}{- the (semi) original list of species provided.}
}
\description{
This function contains all of the logic for passing the data off
to the various webservices that provide APHIAID
}
\seealso{
Other speciesCodes: \code{\link{assignDefinitive}},
  \code{\link{chk_ritis}}, \code{\link{chk_taxize}},
  \code{\link{chk_worrmsTSN}}, \code{\link{chk_worrms}},
  \code{\link{getTSNs}}, \code{\link{getTaxaIDs}},
  \code{\link{skipUselessRecs}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
