% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCodes.R
\name{getCodes}
\alias{getCodes}
\title{getCodes}
\usage{
getCodes(mysteryAPHIAID = NULL, definitiveAPHIAID = NULL,
  APHIAIDs_checked = NULL, mysteryTSN = NULL, definitiveTSN = NULL,
  TSNs_checked = NULL, searchtype = NULL, thisCode = NULL,
  masterlist = NULL, logName = logName)
}
\arguments{
\item{mysteryAPHIAID}{- dataframe with names for which you want to find AphiaIDs}

\item{definitiveAPHIAID}{- dataframe with names for which you already have 
good AphiaIDs.  The results will be bound to this.}

\item{APHIAIDs_checked}{- a vector of APHIAIDs that chave already been sent 
to the webservices (to avoid duplicate checks)}

\item{mysteryTSN}{- dataframe with names for which you want to find TSNs}

\item{definitiveTSN}{- dataframe with names for which you already have 
good TSNs.  The results will be bound to this.}

\item{TSNs_checked}{- a vector of TSNs that chave already been sent 
to the webservices (to avoid duplicate checks)}

\item{searchtype}{- flag indicating whether scientific or common names should 
be used checking the services}

\item{thisCode}{- the particular code that is being looked for (e.g. APHIAID)}

\item{masterlist}{This is the originally submitted list of species (with an 
internally assigned ID)}

\item{logName}{- this is the name of the logfile in the working directory 
that progress should be appended to.}
}
\description{
This function sends names to webservices to try to get codes
}
