% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTSNs.R
\name{getTSNs}
\alias{getTSNs}
\title{getTSNs}
\usage{
getTSNs(mysteryTSN = NULL, doSci = T, doComm = T, knownAphias = NULL,
  masterList = NULL)
}
\arguments{
\item{mysteryTSN}{- df of species for which the TSN is either unknown or not
definitive}

\item{doSci}{- flag indicating whether scientific names should be used when 
checking the services}

\item{doComm}{- flag indicating whether common names should be used when 
checking the services}

\item{knownAphias}{- one worrms check finds TSNs from AphiaIDs}

\item{masterList}{- the (semi) original list of species provided.}
}
\description{
This function contains all of the logic for passing the data off
to the various webservices that provide TSNs.
}
\seealso{
Other speciesCodes: \code{\link{assignDefinitive}},
  \code{\link{chk_ritis}}, \code{\link{chk_taxize}},
  \code{\link{chk_worrmsTSN}}, \code{\link{chk_worrms}},
  \code{\link{getAphiaIDs}}, \code{\link{getTaxaIDs}},
  \code{\link{skipUselessRecs}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
