% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_oracle_cxn.R
\name{make_oracle_cxn}
\alias{make_oracle_cxn}
\title{make_oracle_cxn}
\usage{
make_oracle_cxn(usepkg = "rodbc")
}
\arguments{
\item{usepkg}{default is \code{'rodbc'}. This indicates whether the connection to Oracle should
use \code{'rodbc'} or \code{'roracle'} to connect.  rodbc is slightly easier to setup, but
roracle will extract data ~ 5x faster.}
}
\description{
This function contains the logic for creating a connection to
Oracle.
}
\note{
This is stolen directly from the get_data() function of bio.datawrangling
(also created by me)
# make_oracle_cxn -------------------------------------------------------------
}
\seealso{
Other data: \code{\link{df_to_sp}},
  \code{\link{serviceMaker}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
