% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullDFOGIS.R
\name{pullDFOGIS}
\alias{pullDFOGIS}
\title{pullDFOGIS}
\usage{
pullDFOGIS(service = "ADAPT_Canada_Atlantic_Summer_2016", save_csv = TRUE,
  rec_Start = 0)
}
\arguments{
\item{service}{The default value is
\code{'ADAPT_Canada_Atlantic_Summer_2016'}.  This identifies the service from
which you want to extract data.}

\item{save_csv}{The default value is \code{TRUE}, which means that the
extracted data is saved to a csv file in your working directory.  If
\code{FALSE}, no csv will be created.}

\item{rec_Start}{The default is \code{0}.  Primarily for debugging, this
allows the user to select only some of the records, rather than doing a full
extraction.}
}
\description{
This function extracts an r data.frame from the DFO ESRI REST
service.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
