% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceMaker.r
\name{serviceMaker}
\alias{serviceMaker}
\title{serviceMaker}
\usage{
serviceMaker(channel, write.path)
}
\arguments{
\item{channel}{this is an open RODBC connection to Oracle, with access to GROUNDFISH.FGP_TOWS2 and GROUNDFISH.FGP_TOWS_NW2}

\item{write.path}{This is the path to the geodatabase you will write the data to.}
}
\description{
This function populates a geodatabase with features extracted from the groundfish database.  For every combination of
SEASON/YEAR/SPECIES, a layer is created with all of the null sets
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
