% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skipUselessRecs.R
\name{skipUselessRecs}
\alias{skipUselessRecs}
\title{skipUselessRecs}
\usage{
skipUselessRecs(df = NULL, field = NULL)
}
\arguments{
\item{df}{This is a dataframe of species for which we have no definitive code}

\item{field}{- field in the df containing the value to check against the 
service}
}
\description{
This function is used to prevent records from being sent to
webservices if they do not contain anything that could be used to determine
a definitive code.
}
\seealso{
Other speciesCodes: \code{\link{assignDefinitive}},
  \code{\link{chk_ritis}}, \code{\link{chk_taxize}},
  \code{\link{chk_worrmsTSN}}, \code{\link{chk_worrms}},
  \code{\link{getAphiaIDs}}, \code{\link{getTSNs}},
  \code{\link{getTaxaIDs}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
