% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_ritis.R
\name{chk_ritis}
\alias{chk_ritis}
\title{chk_ritis}
\usage{
chk_ritis(df = NULL, field = NULL, searchtype = NULL)
}
\arguments{
\item{df}{- df of species for which the TSN is either unknown or not
definitive}

\item{field}{- field in the df containing the value to check against the 
service}

\item{searchtype}{- flag indicating whether scientific or common names should 
be used checking the services}
}
\description{
This function sends either the scientific or common name of each
species off to ritis
}
\seealso{
Other speciesCodes: \code{\link{assignDefinitive}},
  \code{\link{chk_taxize}}, \code{\link{chk_worrmsTSN}},
  \code{\link{chk_worrms}}, \code{\link{getAphiaIDs}},
  \code{\link{getTSNs}}, \code{\link{getTaxaIDs}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
