% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxaIDs.R
\name{getTaxaIDs}
\alias{getTaxaIDs}
\title{getTaxaIDs}
\usage{
getTaxaIDs(spec_list = NULL, sci_col = NULL, comm_col = NULL,
  filterStrings = NULL)
}
\arguments{
\item{spec_list}{the dataframe containing information to be decoded to TSN and
aphiaIDs}

\item{sci_col}{the name of the column of the dataframe containing the
scientific names}

\item{comm_col}{the name of the column of the dataframe containing the
common names}

\item{filterStrings}{a vector of regex values that you might want to clean out
of your scientific and common names prior to sending them to a web service.  
For example, some Maritimes names inlude "(NS)", which will prevent services
from finding matches, By adding "\\(NS\\)" (escaping the brackets and periods 
with slashes), we ensure that the results will be as clean as possible prior 
to searching. By default, the following values will be filtered:
\itemize{
\item \\(NS\\)
\item SP\\.
\item S\\.O\\.
\item F\\.
\item UNIDENTIFIED
\item EGGS
\item UNID\\.
\item \\(ORDER\\)
\item \\(SUBORDER\\)
}}
}
\description{
This function attempts to determine a definitive AphiaID and TSN
from various web services via some existing R packages including worrms, 
taxize and ritis.

It takes a species list with Scientific and Common names and performs checks
in the following order:
\enumerate{
  \item Taxize (for AphiaID): Scientific name
  \item Worrms (for AphiaID): Scientific name
  \item Taxize (for AphiaID): Common name
  \item Worrms (for AphiaID): Common name
  \item Worrms (for TSN): <previously found> AphiaIDs
  \item Ritis (for TSN): Scientific name
  \item Ritis (for TSN): Common name
  }
  
In addition to all of the original fields, the returned data
will include: 
\enumerate{
  \item Aphiaid
  \item Aphiaid_src - which service(s) provided the Aphiaid value
  \item Aphiaid_data - which service(s) provided the Aphiaid value
  \item Aphiaid_definitive* - indicates whether or not we are confident about 
  AphiaID
  \item TSN 
  \item TSN_src - which service(s) provided the TSN value
  \item TSN_data - which service(s) provided the TSN value
  \item TSN_definitive* - indicates whether or not we are confident about 
  TSN
  }
Most webservices can return multiple IDs for a given search term, and the 
results are often identified as "accepted" or not. For this function to 
identify an ID as being definitive=TRUE, one of the following 3 situations 
occurred: 
\itemize{
\item Only a single result was returned from a service
\item Only a single item within a resultset was marked as "Accepted"
\item Two independent services both returned the same ID (These may be 
considered less definitive, and can be recognized by their <>_src fields 
indication multiple sources - e.g. "Both worrms & ritis")
}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
